fix: TIO-121: rename localization file
