+++ lib/l10n/app_localizations_extension.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:tiomusic/l10n/en.dart';
-import 'package:tiomusic/l10n/l10n.dart';
+import 'package:tiomusic/l10n/app_localization.dart';
 
 extension AppLocalizationsExtension on BuildContext {
   AppLocalizations get l10n {
