+++ lib/l10n/delegate.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:tiomusic/l10n/de.dart';
 import 'package:tiomusic/l10n/en.dart';
-import 'package:tiomusic/l10n/l10n.dart';
+import 'package:tiomusic/l10n/app_localization.dart';
 
 class AppLocalizationsDelegate extends LocalizationsDelegate<AppLocalizations> {
   @override
