+++ lib/l10n/en.dart
@@ -1,4 +1,4 @@
-import 'package:tiomusic/l10n/l10n.dart';
+import 'package:tiomusic/l10n/app_localization.dart';
 
 class English extends AppLocalizations {
   String get home => 'Home';
